fix for user management

+++ Data Layer/Interface/IUserRepository.cs
@@ -13,9 +13,9 @@ namespace Data_Layer.Interface
         void DisableUser(string userid);
         void EnableUser(string userid);
         void AddUser(IRegisterUser user);
-        IEnumerable<IObjectPair<int, string>> GetRatingCountFromAllUsers();
-        IEnumerable<IObjectPair<int, string>> GetDivergentIABRatingsFromAllUser();
-        IEnumerable<IObjectPair<int, string>> GetDivergentPADRatingsFromAllUser();
+        IEnumerable<IObjectPair<long, string>> GetRatingCountFromAllUsers();
+        IEnumerable<IObjectPair<long, string>> GetDivergentIABRatingsFromAllUser();
+        IEnumerable<IObjectPair<long, string>> GetDivergentPADRatingsFromAllUser();
         void UpdatePassword(string userid, string password, string passwordsalt);
 
     }
