+++ Service Layer/ViewModels/UserManagementViewModel.cs
@@ -15,7 +15,7 @@ namespace Service_Layer.ViewModels
         }
 
         public ISearchUser User { get; set; }
-        public int RatingCount { get; set; }
+        public long RatingCount { get; set; }
         public int ProcentIABDivergent { get; set; }
         public int ProcentPADDivergent { get; set; }
     }
